Split video into submodule, will be a ton of errors
